Added leader election and consensus
Refers to #5

Changed UDPInterface to ComunicationInterface, since it's a better name.

Added ConsensusRole that defines the state of a replica into 3 categories: FOLLOWER, CANDIDATE, LEADER.

RMIServer.java handles most of the consensus mechanism. It isn't a pretty solution, but it works.

ReplicasManagerInterface was added to handle everything related to the replicas address storage. It can be called via RMI if needed.

Implemented missing interfaces on PlaceManager and, added synchronized methods to everything related to replicas memory access.