Swapped comunication protocol between replicas to multicast
Removed Consensus Unicast dependencies, namely ConsensusAppendRequest, ConsensusVoteRequest and ConsensusRequestInterface.

Added new class FullAddress, that contains Address+Port. It is used to reduce complexity. In every part of the code that used to use String+Integer representing Address+Port, was changed to implement this new class.

ComunicationInterface now uses FullAddress.

PlaceManager now uses FullAddress to represent a replica in the ConcurrentHashMap of replicas.

ReplicasManagerInterface now uses FullAddress.

Added new class ComunicationMessage, that contains the heartbeat of the replicas. With the message we want to address, the term of the replica that sends the message and their IP.

ConsensusRole handler now uses FullAddress and ComunicationMessage to comunicate with other replicas.

General code improvements.